Removed packaging, removed operator crate
