+++ Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
 members = [
-    "rust/crd", "rust/operator", "rust/operator-binary"
+    "rust/crd", "rust/operator-binary"
 ]
