--- packaging/debian/stackable-opa-operator.service
@@ -1,13 +1,0 @@
-[Unit]
-Description=Stackable Operator for OpenPolicyAgent
-Before=
-After=network.target
-[Service]
-User=root
-ExecStart=/opt/stackable/opa-operator/stackable-opa-operator
-Restart=on-abort
-StandardOutput=journal
-StandardError=journal
-Environment="OPA_OPERATOR_LOG=info"
-[Install]
-WantedBy=multi-user.target
