+++ rust/operator-binary/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-authors = ["Malte Sander <malte.sander.it@gmail.com>"]
+authors = ["Stackable GmbH <info@stackable.de>"]
 build = "build.rs"
 description = "Stackable Operator for OpenPolicyAgent"
 edition = "2018"
@@ -9,31 +9,27 @@ repository = "https://github.com/stackabletech/opa-operator"
 version = "0.7.0-nightly"
 
 [dependencies]
+failure = "0.1"
+futures = { version = "0.3", features = ["compat"] }
+semver = "1.0"
+serde = "1.0"
+serde_json = "1.0"
+serde_yaml = "0.8"
+snafu = "0.6"
+structopt = "0.3"
+strum = "0.22"
+strum_macros = "0.22"
+tokio = { version = "1.12", features = ["full"] }
+tokio01 = { version = "0.1", package = "tokio" }
+tokio-executor = "0.1"
+tracing = "0.1"
+pin-project = "1.0"
 stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.6.0" }
 stackable-opa-crd = { path = "../crd" }
-stackable-opa-operator = { path = "../operator" }
-
-clap = "2.33"
-serde_yaml = "0.8"
-tokio = { version = "1.13", features = ["macros", "rt-multi-thread"] }
-tracing = "0.1"
-structopt = "0.3.25"
+anyhow = "1.0"
+fnv = "1.0"
 
 [build-dependencies]
 built = { version =  "0.5", features = ["chrono", "git2"] }
 stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.6.0" }
 stackable-opa-crd = { path = "../crd" }
-
-[package.metadata.deb]
-name = "stackable-opa-operator"
-maintainer-scripts = "../../packaging/debian/"
-systemd-units = { enable = false, unit-name = "stackable-opa-operator" }
-assets = [
-    ["../../target/release/stackable-opa-operator", "opt/stackable/opa-operator/", "755"],
-    ["../../deploy/crd/openpolicyagent.crd.yaml", "etc/stackable/opa-operator/crd/", "644"],
-    ["../../deploy/config-spec/properties.yaml", "etc/stackable/opa-operator/config-spec/", "644"],
-]
-
-[[bin]]
-name = "stackable-opa-operator"
-path = "src/stackable-opa-operator.rs"
