--- rust/operator/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-authors = ["Malte Sander <malte.sander.it@gmail.com>"]
-description = "Stackable Operator for OpenPolicyAgent - The Operator Library"
-edition = "2018"
-license = "OSL-3.0"
-name = "stackable-opa-operator"
-repository = "https://github.com/stackabletech/opa-operator"
-version = "0.7.0-nightly"
-
-[dependencies]
-stackable-opa-crd = { path = "../crd" }
-stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.6.0" }
-
-async-trait = "0.1"
-futures = "0.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-strum = "0.22"
-strum_macros = "0.22"
-thiserror = "1.0"
-tokio = { version = "1.13", features = ["macros"] }
-tracing = "0.1"
-snafu = "0.6.10"
-
-[features]
